proposal: switch bool default value
Even if both is working, I think that defining the actual `bool` default value NOT as a string might be less confusing. I just stumbled upon this one and thought: "Ha, why is that a 'string'?". Just my opinion and I thought that you might be interested in my thought process :) Feel free to close this PR if you want to stick with the string definition :)