fix(flow): export prisma interface
Fixes https://github.com/prisma/prisma/issues/3630

This makes exporting same as TS. Why was it originally not exported
though?

Renaming it as suggested by https://github.com/prisma/prisma/issues/3630
is not needed as we are exporting a constant by the name "Prisma" (same as TS).

Tested this client with the script example:
https://github.com/prisma/prisma-examples/tree/master/flow/script

Both "yarn flow" and "yarn start" commands work as expected.