Increasing test-server startup delay
This change is required because of two things:

1. There are now several parse-attempts of multiple datamodel files.
   This just generally adds overhead when starting a prisma server
2. The code is being run in `Debug` mode which generally is a good
   idea, but in this case means startup is much slower.

Ideally we want to change the structure of how tests are run anyway
but for now this change should do. It's just long enough for the
server to initialise in most cases. Do be aware that this number might
have to be adjusted in the future.