First pass at a "look ahead" parser implementation
The idea here is the following: if we have any fields in a mutation
that are required, but would only be filled-in by a future mutation
(i.e. a nested connect), we fold it into the level above.

This state currently has some limitations that need to be expanded on:

- Only handles create -> connect case
- Only handles one nesting level-cases (i.e. top-level -> connect)
- Arbitrary nested or different actions ignored