Very work in progress
The idea behind this change is to restructure the LookAhead module to
be able to yield mixed-type mutation trees. This means that a
dependent mutation can have ReadQuery's in it, that need to be
executed in order to make other WriteQuery's work.

Things become a bit more complicated when also considering nested
operations, i.e. any nested operation can result in a new root-tree,
with mixed Read and Write operations.